Changed print_to_hex and some minor things

+++ loader/include/io.h
@@ -1,9 +1,10 @@
 #ifndef IO_H
 #define IO_H
+#include <stdint.h>
 
 void print_str(char * str);
-void int_to_hex(char * buffer, long long n, char upper);
-void print_hex(long long i);
+void int_to_hex(char * buffer, uint64_t n, char upper);
+void print_hex(uint64_t i);
 void printf(const char* format, ...);
